CLN: rename iterate_images as analysis_generator, and move it into top level package
